requirements.txt
discord.py
pyNacl
cachetools
pytz
